+++ src/addon.ts
@@ -7,7 +7,6 @@ class Addon {
         alive: boolean;
         // Env type, see build.js
         env: "development" | "production";
-        // ztoolkit: MyToolkit;
         ztoolkit: ZToolkit;
         locale?: {
             current: any;
@@ -30,7 +29,6 @@ class Addon {
         this.data = {
             alive: true,
             env: __env__,
-            // ztoolkit: new MyToolkit(),
             ztoolkit: createZToolkit(),
             dialogs: {},
         };
@@ -38,51 +36,5 @@ class Addon {
         this.api = {};
     }
 }
-
-/**
- * Alternatively, import toolkit modules you use to minify the plugin size.
- *
- * Steps to replace the default `ztoolkit: ZoteroToolkit` with your `ztoolkit: MyToolkit`:
- *
- * 1. Uncomment this file's line 30:            `ztoolkit: new MyToolkit(),`
- *    and comment line 31:                      `ztoolkit: new ZoteroToolkit(),`.
- * 2. Uncomment this file's line 10:            `ztoolkit: MyToolkit;` in this file
- *    and comment line 11:                      `ztoolkit: ZoteroToolkit;`.
- * 3. Uncomment `../typings/global.d.ts` line 12: `declare const ztoolkit: import("../src/addon").MyToolkit;`
- *    and comment line 13:                      `declare const ztoolkit: import("zotero-plugin-toolkit").ZoteroToolkit;`.
- *
- * You can now add the modules under the `MyToolkit` class.
- */
-
-// import { BasicTool, unregister } from "zotero-plugin-toolkit/dist/basic";
-// import { UITool } from "zotero-plugin-toolkit/dist/tools/ui";
-// import { PreferencePaneManager } from "zotero-plugin-toolkit/dist/managers/preferencePane";
-// import { ShortcutManager } from "zotero-plugin-toolkit/dist/managers/shortcut";
-// import { MenuManager } from "zotero-plugin-toolkit/dist/managers/menu";
-// import { FieldHookManager } from "zotero-plugin-toolkit/dist/managers/fieldHook";
-// import { ProgressWindowHelper } from "zotero-plugin-toolkit/dist/helpers/progressWindow";
-
-// export class MyToolkit extends BasicTool {
-//   UI: UITool;
-//   PreferencePane: PreferencePaneManager;
-//   Shoutcut: ShortcutManager;
-//   Menu: MenuManager;
-//   FieldHooks: FieldHookManager;
-//   // ProgressWindow: typeof ProgressWindowHelper;
-
-//   constructor() {
-//     super();
-//     this.UI = new UITool(this);
-//     this.PreferencePane = new PreferencePaneManager(this);
-//     this.Shoutcut = new ShortcutManager(this);
-//     this.Menu = new MenuManager(this);
-//     this.FieldHooks = new FieldHookManager(this);
-//     // this.ProgressWindow = new ProgressWindowHelper(this);
-//   }
-
-//   unregisterAll() {
-//     unregister(this);
-//   }
-// }
 
 export default Addon;