+++ src/modules/preference.ts
@@ -3,13 +3,12 @@ import { getString } from "../utils/locale";
 import { getPref, setPref } from "../utils/prefs";
 
 export function registerPrefs() {
-    ztoolkit.PreferencePane.register({
+    Zotero.PreferencePanes.register({
         pluginID: config.addonID,
         src: rootURI + "chrome/content/preferences.xhtml",
         label: getString("prefs-title"),
         image: `chrome://${config.addonRef}/content/icons/favicon.png`,
         helpURL: homepage,
-        defaultXUL: true,
     });
 }