+++ src/utils/ztoolkit.ts
@@ -1,19 +1,25 @@
 // import ZoteroToolkit from "../../../zotero-plugin-toolkit/dist/index";
 import { config } from "../../package.json";
-import ZoteroToolkit from "zotero-plugin-toolkit";
+// import ZoteroToolkit from "zotero-plugin-toolkit";
 import { BasicTool, unregister } from "zotero-plugin-toolkit/dist/basic";
-import { PreferencePaneManager } from "zotero-plugin-toolkit/dist/managers/preferencePane";
+import { DialogHelper } from "zotero-plugin-toolkit/dist/helpers/dialog";
+import { FilePickerHelper } from "zotero-plugin-toolkit/dist/helpers/filePicker";
+import { ProgressWindowHelper } from "zotero-plugin-toolkit/dist/helpers/progressWindow";
+import { VirtualizedTableHelper } from "zotero-plugin-toolkit/dist/helpers/virtualizedTable";
+import { MenuManager } from "zotero-plugin-toolkit/dist/managers/menu";
+import { ShortcutManager } from "zotero-plugin-toolkit/dist/managers/shortcut";
+import { ExtraFieldTool } from "zotero-plugin-toolkit/dist/tools/extraField";
 import { UITool } from "zotero-plugin-toolkit/dist/tools/ui";
 
 export { createZToolkit };
 
 function createZToolkit() {
-    const _ztoolkit = new ZoteroToolkit();
+    // const _ztoolkit = new ZoteroToolkit();
     /**
      * Alternatively, import toolkit modules you use to minify the plugin size.
      * You can add the modules under the `MyToolkit` class below and uncomment the following line.
      */
-    // const _ztoolkit = new MyToolkit();
+    const _ztoolkit = new MyToolkit();
     initZToolkit(_ztoolkit);
     return _ztoolkit;
 }
@@ -24,6 +30,7 @@ function initZToolkit(_ztoolkit: ReturnType<typeof createZToolkit>) {
     _ztoolkit.basicOptions.log.disableConsole = env === "production";
     _ztoolkit.UI.basicOptions.ui.enableElementJSONLog = __env__ === "development";
     _ztoolkit.UI.basicOptions.ui.enableElementDOMLog = __env__ === "development";
+    _ztoolkit.UI.basicOptions.ui.enableElementRecord = __env__ === "development";
     _ztoolkit.basicOptions.debug.disableDebugBridgePassword = __env__ === "development";
     _ztoolkit.basicOptions.api.pluginID = config.addonID;
     _ztoolkit.ProgressWindow.setIconURI("default", `chrome://${config.addonRef}/content/icons/favicon.png`);
@@ -31,12 +38,24 @@ function initZToolkit(_ztoolkit: ReturnType<typeof createZToolkit>) {
 
 class MyToolkit extends BasicTool {
     UI: UITool;
-    PreferencePane: PreferencePaneManager;
+    Menu: MenuManager;
+    Shortcut: ShortcutManager;
+    ProgressWindow: typeof ProgressWindowHelper;
+    VirtualizedTable: typeof VirtualizedTableHelper;
+    Dialog: typeof DialogHelper;
+    FilePicker: typeof FilePickerHelper;
+    ExtraField: ExtraFieldTool;
 
     constructor() {
         super();
         this.UI = new UITool(this);
-        this.PreferencePane = new PreferencePaneManager(this);
+        this.Menu = new MenuManager(this);
+        this.Shortcut = new ShortcutManager(this);
+        this.ProgressWindow = ProgressWindowHelper;
+        this.VirtualizedTable = VirtualizedTableHelper;
+        this.Dialog = DialogHelper;
+        this.FilePicker = FilePickerHelper;
+        this.ExtraField = new ExtraFieldTool(this);
     }
 
     unregisterAll() {